Scenario 72: Oozing Grove
The party venture into the Oozing grove to placate a Councilman, for Dominic the Town Records writer.
They are tasked with clearing it, destroying the fruit trees creating oozes.
Initially the Sawbones takes heavy damage near the middle tree. The team stabilise after some executions from the Nightshroud and a huge volley of axes from the Berserker, but Oozes begin amassing at one end. They clear two trees and prepare, before advancing towards the other side. The Sunkeeper and Berserker advance, but are hard hit by the oozes unexpectedly attacking both of them. The Berkserker swings her axe and takes down two before collapsing exhausted.
The remaining team cut through their way to the final tree. The Sawbones makes a valiant last stand, disarming and executing some oozes.
The Sunkeeper and Nighthshroud, and unlikely pairing, were both tired, and faced with whittling down the tree or trying to clear the final oozes. They finished the tree, but the oozes split, and the Sunkeeper collapsed.
Facing down a healthy ooze, the Nightshroud thought the final task was impossible. He hit the ooze as hard as possible, but it split, making the whole drawn out affair seem futile. As a last stand, he cloaked himself and prayed for the only hope left. The oozes obliged, and split once more, killing themselves, a moment before the Nightshroud was about to fall.