Sencario 35: Gloomhaven Battlements A
The party return to the Prime Demon in Gloomhaven with the Artifact, and aid him in taking over the city and instilling demonic rule.
With their demonic allies they smash through the River Gate. The Berserker charges in and throws axes at all in range, instantly killing many, and wounding the rest, before rushing at the Captain. The party makes short work of the captain and his remaining allies, and return his head to the Prime Demon.