Scenario 25: Icecrag ascent
The party returned to the mountain in search of a dragon. The Sawbones was particularly interested in finding Spitting Drakes for his poison concoction.
The Sunkeeper tanked the assaults by the hounds as the others dispatched them. The Sawbones and Nighthshroud executed the drakes in the next room, as the Berserker, having completed her quest, prepared for a last stand in the final room. This didn't go quite to plan, but with the help of the Nightshroud they cleared the remaining drakes and made it to the moutain's summit, meeting the fabled drake.
Upon returning to Gloomhaven, the Berserker, having killed enough strong enemies, retires.