Scenario 21: Infernal Throne
After aiding the Prime Demon, the party, aided by their new Beast Tyrant ally, head towards the Rift where they first encountered the Prime Demon.
They want their artifact back, and Esper the Nightshroud has Demons to hunt.
They make short work of the first demons encountered. The Sawbones grows impatient, and rushes through to the next area, before realising his mistake and scurrying back.
After clearing the room, they decide to head North, to clear the area before approaching the Prime Demon. The Sunkeeper realises how long this will take her, and changes tack, heading for the Prime Demon's lair. The Beast Tyrant holds off the Demons to the North, while the others make short work of the Prime Demon's altar.
Upon realising they will soon be victorious, the Nightshroud makes for the chest, and the Beast Tyrant exchanges places with her to grab it. The Sawbones uses his speed to beat the Tyrant to the chest, but allows the Tyrant to take it, in exchange for help tracking down the Lurkers for his poison.