Scenario 93: Sunken Vessel
The party goes in search of treasure and find a sunken ship and deserted island, where they are ambushed by lurkers.

They make quick work of the first lurkers, but then make glacial progress through the next narrow corridor. The bear takes the opportunity to fill its boots with loot.
The newly joined Plagueherald and Tetra the Elementalist exhaust in the final room, but the others take their time to grab as much loot as they can.