Scenario 33: Savvas Armoury
The party enter the Savvas Armoury, searching for a scale for the Drake, and a vessel for the Voice.
The Savvas were formidable foes, but stunning them before they could complete their summoning rituals proved an extremely effective strategy.
After treading on the pressure plates, the team were ambushed by reinforcements. They proceeded into the next room, but the bear had other ideas, and ran back to deal with the ambush.
Eventually after dealing with the threats, Tetra was teleported back to the entrance and the gang made their escape