Scenario 45: Rebel Swamp
The party head off to put down the rebels fighting against the demons. It starts poorly, the guards and archers target the Sunkeeper and land some hugely damaging blows.
After taking heavy damage the final archer deals a near-fatal blow to the Beast Tyrant. The Sunkeeper comes to his aid to prevent his exhaustion.
After destroying many totems the Plagueherald makes a mad dash to finish the final one, as they are struggling to kill the final guard.
A dying guard tells them where to find the final rebels, but also beseeches them to take down the demons instead. Regime change is a favourite hobby of theirs...