Scenario 6: Decaying crypt
Returning to a crypt they discovered long ago, the party
They are met with a horde of renaimated skeletons and corpses in the first room, but between their disables and tanking, manage to clear it without any real casualties.
In the second tunnel room, Tetra the Elementalist rushes forward to obliterate some living spirits before they prove too much of a nuisance, and the skeletons are mostly held at bay, despite getting some surprisingly quick hits in on the Sunkeeper.
The final room does not prove too much of a challenge, and the party spends much of their time leisurely grabbing gold.

They meet a robed figure who instructs them to kill Jekserah. Apparently he's been in this crypt a while, as they'd helepd Jekserah take over the city, then changed the regime twice over since...