Scenarios 59 & 60: The cure for the Elementalist
The party are misled, while searching for a cure a Quatryl sends them into an ambush to die, but they make short work of it.

They return to find her running from a burning building, and charge in to rescue the cure.
Lyra and Rocky, the faster members of the party, charge to the back, while Tetra and the Doomstalker open a door into a room full of snakes. Somehow they manage, with wide ranging attacks, to clear them all out as the roof starts caving in. Lyra and Rocky make short work of the drakes, and the Cragheart crushes the hounds in the second room, while Lyra focuses on slowly picking up coins, as her pet summons are crushed by the building.
Tetra takes a lot of hits from the Oozes, but with some help finishes them off, and hobbles back to the entrance, as the others grab the last few things.